+++ systems/trajectory_optimization/dircon/BUILD.bazel
@@ -1,0 +1,18 @@
+# -*- mode: python -*-
+# vi: set ft=python :
+
+package(default_visibility = ["//visibility:public"])
+
+cc_library(
+    name = "dircon",
+    srcs = [
+        "dircon_mode.cc",
+    ],
+    hdrs = [
+        "dircon_mode.h",
+    ],
+    deps = [
+        "//multibody/kinematic",
+        "@drake//:drake_shared_library",
+    ],
+)
