+++ systems/trajectory_optimization/dircon/dircon_mode.h
@@ -1,6 +1,5 @@
 #pragma once
-#include <unordered_map>
-#include <vector>
+#include <map>
 #include <set>
 
 #include "multibody/kinematic/kinematic_evaluator_set.h"
